Code/requirements.txt
pandas==1.4.2
numpy>=1.21.4
scipy>=1.7.2
avalanche-lib==0.1.0
torch>=1.8.1
torchvision==0.11.3
scikit-learn==1.0.2
Pillow==9.0.1